feat(mobile): iPad pane follows selection; wall gets its own surfaces
The iPad shell's right pane mirrored the live wall, which collapsed shared
status onto the selection surface — the pane read as ambiguous and tapping a
climb did nothing (browse→inspect was dead). Make the pane follow selection
(master-detail) and give the live wall its own labelled, width-adaptive home.

- PlayDrawer: pane shows currentClimbQueueItem with real prev/next (drop the
  isPane→wallClimb override + nav short-circuit); add an "On the wall" chip when
  the selection is the lit climb; add paneTopInset so a docked strip owns it.
- Wall surfaces escalate by a pure width budget (resolveWallSurface in
  size-class.ts): ambient SidebarWallCell (rail footer), WallStrip atop the pane
  in portrait, dedicated IpadWallColumn in landscape — all reusing
  NowOnTheWallPanel, extracted from BoardSheet (now a thin modal wrapper).
- drawer-host exposes boardPanelProps (mirrors playDrawerPaneProps); column
  gated on a resolved active board so it never reserves dead space.
- i18n: boardPresence.railDark; playView.paneEmpty re-pointed to selection.

Phone (compact) path unchanged.
