+++ packages/shared/i18n/locales/es/session.json
@@ -26,6 +26,7 @@
   "boardPresence": {
     "title": "Ahora en el muro",
     "open": "En el muro",
+    "railDark": "Nada encendido ahora",
     "openAria": "Ver qué hay en el muro",
     "openAriaWithClimb": "Ver qué hay en el muro: {{name}}",
     "nowOnWallAnnouncement": "Ahora en el muro: {{name}}",
@@ -486,8 +487,8 @@
     "previewBadge": "Vista previa",
     "setActive": "Activar",
     "paneEmpty": {
-      "title": "Nada en el muro",
-      "subtitle": "El bloque encendido aparece aquí."
+      "title": "Elige un bloque",
+      "subtitle": "Toca un bloque para verlo aquí."
     },
     "previousFrame": "Fotograma anterior",
     "nextFrame": "Fotograma siguiente",
